+++ console/components/messages/Embed.jsx
@@ -1,0 +1,3 @@
+export default function Embed({}) {
+  return <div></div>
+}
