+++ console/services/app/App.ts
@@ -1,12 +1,5 @@
 import Provider from './Provider'
 
-// /channel/apps/add asana <api-token>
-// /channel/apps/add asana <api-token> as:client-asana
-// /close https://app.asana.com/0/1138929626133616/1177203200389571
-// /comment https://app.asana.com/0/1138929626133616/1177203200389571 "done!"
-// for task that is only in client-asana
-// /client-asana/close https://app.asana.com/0/1138143243423432/1123439392834213323
-
 export interface AppSpec {
   name: string
   defaultProvider?: string
