+++ pages/embed.js
@@ -1,0 +1,3 @@
+export default ({}) => {
+  return <h1>This is an embed</h1>
+}
